# Conflicts:
#	TradeSystem-Bundle/pom.xml
#	TradeSystem-Packets/pom.xml
#	TradeSystem-Spigot/pom.xml
#	pom.xml

+++ TradeSystem-Spigot/src/main/java/de/codingair/tradesystem/spigot/trade/layout/types/impl/economy/EconomyIcon.java
@@ -193,7 +193,7 @@ public abstract class EconomyIcon<T extends Transition.Consumer<Double> & TradeI
 
     @Override
     public void deserialize(@NotNull DataInputStream in) throws IOException {
-        this.value = in.readInt();
+        this.value = in.readDouble();
     }
 
     @Override
